Restructuring project files
