+++ app/src/main/java/jcuratcha/weather/Clients/WeatherInformationClient.java
@@ -1,7 +1,4 @@
-package jcuratcha.weather.Clients;
-
-import com.android.volley.RequestQueue;
-import com.android.volley.toolbox.Volley;
+package jcuratcha.weather.clients;
 
 /**
  * Created by DoorCrasher on 2016-12-07.
