+++ app/src/main/java/jcuratcha/weather/WebClients/CurrentWeatherClient.java
@@ -1,4 +1,4 @@
-package jcuratcha.weather.WebClients;
+package jcuratcha.weather.webclients;
 
 import android.app.Activity;
 import android.util.Log;
@@ -6,20 +6,16 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.android.volley.Request;
-import com.android.volley.RequestQueue;
 import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.JsonObjectRequest;
-import com.android.volley.toolbox.Volley;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.util.Locale;
 
-import jcuratcha.weather.Dashboard;
-import jcuratcha.weather.Objects.CurrentWeatherObject;
-import jcuratcha.weather.RequestQueueSingleton;
+import jcuratcha.weather.objects.CurrentWeatherObject;
 
 /**
  * CurrentWeatherClient
