+++ app/src/main/res/layout/activity_main.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="jcuratcha.weather.Dashboard">
+    tools:context="jcuratcha.weather.activities.MainActivity">
     <LinearLayout
         android:orientation="vertical"
         android:layout_width="match_parent"
